Bugfix: White space error when converting to ass
When parsing SMI into HTML, I've made a decision to preserve the next line character since I've snow that some subtitle creators are using the next line character as a line subtitle (this is not an official way of doing it). In the initial testing it was seems working fine, but I just found that this "next line character" break format of ASS that video player was not reading subtitle correctly.

So, code has been updated to just remove all of next line character when converting subtitle line.